requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
build==1.4.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chromadb==1.1.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
commoncode==32.2.0
contourpy==1.3.0
crewai==1.11.0
cryptography==44.0.1
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.53.1
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.8
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.14.5
ipykernel==6.30.1
ipython==8.37.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
kubernetes==35.0.0
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lancedb==0.30.0
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-core==0.3.69
langchain-google-genai==2.0.10
langchain-huggingface==0.3.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8
lark==1.2.2
linkify-it-py==2.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
narwhals==2.18.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.2
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.29.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.0
overrides==7.7.0
packaging==24.1
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.5
pdfminer.six==20251230
pdfplumber==0.11.9
peewee==4.0.2
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.5.0
plugincode==32.0.0
portalocker==2.7.0
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.22
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.1.4
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.22
pytz==2023.3.post1
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.5.3
saneyaml==0.6.1
scikit-learn==1.5.1
scipy==1.14.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
sse-starlette==3.3.3
stack-data==0.6.3
starlette==0.52.1
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
textual==8.1.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.0.2
tomli_w==1.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typecode==30.0.2
typer==0.23.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2023.3
uc-micro-py==2.0.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.2
uv==0.9.30
uvicorn==0.42.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
yarl==1.23.0
yfinance==1.2.0
zipp==3.23.0
zstandard==0.23.0
